+++ scripts/R/pulser/utils.R
@@ -36,7 +36,7 @@ whichHigh <- function(x, level) {
 
 
 ##' Set all options
-setOpts <- function(bounds, tolerance, cores = 24, replicates = 5, normFactors = NULL) {
+setOpts <- function(bounds, tolerance, cores = 40, replicates = 5, normFactors = NULL) {
   opts <- setFittingOptions(verbose = "verbose")
   opts$cores <- cores
   opts$replicates <- replicates